refactor: 관심가게 관련 api 리팩토링 및 토큰 생성 수정

Closes S12P11D102-281

See merge request s12-webmobile1-sub1/S12P11D102!76

+++ backend/bbanggu/src/main/java/com/ssafy/bbanggu/favorite/FavoriteService.java
@@ -1,63 +1,95 @@
 package com.ssafy.bbanggu.favorite;
 
-import com.ssafy.bbanggu.bakery.Bakery;
-import com.ssafy.bbanggu.user.domain.User;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.stream.Collectors;
+
+import com.ssafy.bbanggu.bakery.Bakery;
+import com.ssafy.bbanggu.bakery.BakeryRepository;
+import com.ssafy.bbanggu.bakery.dto.BakeryDetailDto;
+import com.ssafy.bbanggu.bakery.service.BakeryService;
+import com.ssafy.bbanggu.common.exception.CustomException;
+import com.ssafy.bbanggu.common.exception.ErrorCode;
+import com.ssafy.bbanggu.user.domain.User;
+import com.ssafy.bbanggu.user.repository.UserRepository;
+
+import lombok.RequiredArgsConstructor;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class FavoriteService {
 
 	private final FavoriteRepository favoriteRepository;
+	private final BakeryService bakeryService;
+	private final UserRepository userRepository;
+	private final BakeryRepository bakeryRepository;
 
 	// 좋아요 추가
 	@Transactional
-	public String addFavorite(Long userId, Long bakeryId) {
+	public void addFavorite(Long userId, Long bakeryId) {
+		User user = userRepository.findById(userId)
+			.orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+		Bakery bakery = bakeryRepository.findById(bakeryId)
+			.orElseThrow(() -> new CustomException(ErrorCode.BAKERY_NOT_FOUND));
+
 		// 이미 좋아요한 상태인지 확인
 		if (favoriteRepository.existsByUser_UserIdAndBakery_BakeryId(userId, bakeryId)) {
-			throw new IllegalArgumentException("이미 좋아요한 가게입니다.");
+			throw new CustomException(ErrorCode.ALREADY_FAVORITE);
 		}
-
-
-		User user = User.builder()
-			.userId(userId)
-			.build();
-
-		Bakery bakery = Bakery.builder()
-			.bakeryId(bakeryId)
-			.build();
 
 		// 새로 좋아요 추가
 		Favorite favorite = Favorite.builder()
 			.user(user)
 			.bakery(bakery)
-			.isLiked(true)
 			.build();
 
 		favoriteRepository.save(favorite);
-		return "좋아요가 등록되었습니다.";
 	}
 
 	// 좋아요 취소
 	@Transactional
-	public String removeFavorite(Long userId, Long bakeryId) {
+	public void removeFavorite(Long userId, Long bakeryId) {
+		if (!userRepository.existsById(userId)) {
+			throw new CustomException(ErrorCode.USER_NOT_FOUND);
+		}
+
+		if (!bakeryRepository.existsById(bakeryId)) {
+			throw new CustomException(ErrorCode.BAKERY_NOT_FOUND);
+		}
+
 		Favorite favorite = favoriteRepository.findByUser_UserIdAndBakery_BakeryId(userId, bakeryId)
-			.orElseThrow(() -> new IllegalArgumentException("좋아요 하지 않은 가게입니다."));
+			.orElseThrow(() -> new CustomException(ErrorCode.BAKERY_NOT_FAVORITE));
 
 		favoriteRepository.delete(favorite);
-		return "좋아요가 취소되었습니다.";
 	}
 
 	// 유저가 좋아요한 모든 가게 조회
 	@Transactional(readOnly = true)
-	public List<Bakery> findAllFavorites(Long userId) {
-		return favoriteRepository.findByUser_UserId(userId).stream()
-			.map(Favorite::getBakery)
-			.collect(Collectors.toList());
+	public Page<BakeryDetailDto> findAllFavorites(Long userId, double userLat, double userLng, Pageable pageable) {
+		if (!userRepository.existsById(userId)) {
+			throw new CustomException(ErrorCode.USER_NOT_FOUND);
+		}
+
+		// 좋아요 누른 가게 아이디 조회
+		List<Favorite> favorites = favoriteRepository.findByUser_UserId(userId);
+		List<Long> bakeryIds = favorites.stream()
+			.map(f -> f.getBakery().getBakeryId())
+			.toList();
+
+		if(bakeryIds.isEmpty()) return Page.empty();
+
+		Page<Bakery> bakeries = bakeryRepository.findByBakeryIdInAndDeletedAtIsNull(bakeryIds, pageable);
+
+		List<BakeryDetailDto> bakeryDetailDtos = bakeries.getContent().stream()
+			.map(b -> BakeryDetailDto.from(b, bakeryService.calculateDistance(userLat, userLng, b.getLatitude(), b.getLongitude())))
+			.toList();
+
+		return new PageImpl<>(bakeryDetailDtos, pageable, bakeries.getTotalElements());
 	}
 }
